Memory optimisations following Balmoral problem case investigation
Avoid calling realloc if possible. So don't trim size of any returned matrices, which is wasteful. Add new cs_lu_mem that accepts L and U and so can re-use matrices that were allocated previously. For example in an earlier call to cs_lu_mem. Experimentation shows that this eliminates a lot of allocation overhead.

These changes are especially important for huge allocations which will map to direct VirtualAlloc calls which involve kernel mode transition and system wide lock contention.